+++ changelog.md
@@ -3,6 +3,7 @@
 ### 1.8.1 (unreleased)
 
 * Don't require Uberalls to be enabled to post coverage data to Harbormaster
+* Handle UTF-8 strings properly in comment file
 
 ### 1.8.0 (2015/09/09)